+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Support parsing name from Poetry 2.0 pyproject.toml
 * Handle bad UTF-8 in METADATA files
+* Experimental support for converting binary (non-pure Python) wheels via `--allow-impure` flag
 
 ### Bug fixes